requirements copy.txt
Flask==2.1.0
tensorflow==2.8.0
numpy==1.24.2
opencv-python-headless==4.7.0.68
Pillow